Fix group case
Make group title-case, because Jitpack is case-sensetive and uses `group=com.github.<userame>`, where `<username>` has the case from Github (which is TitleCase (Lipen) for me). Lower-case query for Jitpack forces it to rebuild the package for different coordinates (group), so it is better to stick to the correct case (implied by Github username) everywhere.